+++ plugins/view-resources/src/components/filter/ObjectFilter.svelte
@@ -16,7 +16,7 @@
   import { Class, Doc, FindResult, getObjectValue, Ref, RefTo, SortingOrder } from '@anticrm/core'
   import { translate } from '@anticrm/platform'
   import presentation, { getClient } from '@anticrm/presentation'
-  import ui, { Button, CheckBox, Label, Loading } from '@anticrm/ui'
+  import ui, { Button, CheckBox, eventToHTMLElement, Label, Loading, showPopup } from '@anticrm/ui'
   import { Filter } from '@anticrm/view'
   import { onMount } from 'svelte'
   import { buildConfigLookup, getPresenter } from '../../utils'
@@ -24,6 +24,8 @@
   import { createEventDispatcher } from 'svelte'
   import task from '@anticrm/task'
   import type { State } from '@anticrm/task'
+  import FilterTypePopup from './FilterTypePopup.svelte'
+  import { FilterQuery } from '../../filter'
 
   export let _class: Ref<Class<Doc>>
   export let filter: Filter
@@ -35,13 +37,13 @@
   const key = { key: tkey }
   const lookup = buildConfigLookup(hierarchy, _class, [tkey])
   const promise = getPresenter(client, _class, key, key, lookup)
-  filter.modes = [view.ids.FilterObjectIn, view.ids.FilterObjectNin]
+  filter.modes = filter.modes === undefined ? [view.filter.FilterObjectIn, view.filter.FilterObjectNin] : filter.modes
   filter.mode = filter.mode === undefined ? filter.modes[0] : filter.mode
 
   let values: (Doc | undefined | null)[] = []
   let objectsPromise: Promise<FindResult<Doc>> | undefined
-  const targetClass = (hierarchy.getAttribute(_class, filter.key.key).type as RefTo<Doc>).to
-  const clazz = hierarchy.getClass(targetClass)
+  $: targetClass = (hierarchy.getAttribute(_class, filter.key.key).type as RefTo<Doc>).to
+  $: clazz = hierarchy.getClass(targetClass)
   const targets = new Map<any, number>()
   $: isState = clazz._id === task.class.State ?? false
   let statesCount: number[] = []
@@ -134,9 +136,42 @@
 
   const dispatch = createEventDispatcher()
   getValues(search)
+
+  $: byCriteria = hierarchy.hasMixin(clazz, view.mixin.ClassFilters)
+
+  function setNestedFilter (e: Filter | undefined) {
+    if (e === undefined) return
+    filter.nested = e
+    filter.mode = view.filter.FilterNestedMatch
+    filter.modes = [view.filter.FilterNestedMatch, view.filter.FilterNestedDontMatch]
+    filter.value = e.value
+    filter.onRemove = () => {
+      FilterQuery.remove(filter.index)
+    }
+    onChange(filter)
+    dispatch('close')
+  }
+
+  function nestedFilter (e: MouseEvent) {
+    const target = eventToHTMLElement(e)
+    showPopup(
+      FilterTypePopup,
+      {
+        _class: targetClass,
+        target,
+        index: filter.index,
+        filter: filter.nested,
+        onChange: setNestedFilter
+      },
+      target
+    )
+  }
 </script>
 
 <div class="selectPopup">
+  {#if byCriteria}
+    <Button shape={'round'} label={view.string.MatchCriteria} on:click={nestedFilter} />
+  {/if}
   {#if clazz.sortingKey}
     <div class="header">
       <input